Rename App to AppComponent to resolve import conflict with app.js
