+++ frontend/lint/module-boundaries.mjs
@@ -147,7 +147,7 @@ const elements = [
     "frontend/src/metabase/app-main.js",
     "frontend/src/metabase/app-embed.ts",
     "frontend/src/metabase/app-public.ts",
-    "frontend/src/metabase/App.tsx",
+    "frontend/src/metabase/AppComponent.tsx",
     "frontend/src/metabase/App.styled.tsx",
     "frontend/src/metabase/AppKBarProvider.tsx",
     "frontend/src/metabase/reducers-main.ts",
