+++ frontend/src/metabase/AppComponent.tsx
