+++ frontend/src/metabase/routes.tsx
@@ -1,7 +1,7 @@
 import type { Store, ThunkDispatch, UnknownAction } from "@reduxjs/toolkit";
 import { IndexRedirect, IndexRoute, Redirect, Route } from "react-router";
 
-import App from "metabase/App";
+import App from "metabase/AppComponent";
 import { getAccountRoutes } from "metabase/account/routes";
 import CollectionPermissionsModal from "metabase/admin/permissions/components/CollectionPermissionsModal/CollectionPermissionsModal";
 import { getRoutes as getAdminRoutes } from "metabase/admin/routes";
